Move func wavesurfer_play to manymusic-annotator
This prevents manymusic-annotator from importing the utils module,
which allows to avoid importing development dependencies (e.g.,
scikitlearn, matplotlib...), allowing for a lightweight annotation
environment.
