+++ manymusic-annotator.py
@@ -8,7 +8,6 @@ import streamlit as st
 import streamlit.components.v1 as components
 import pandas as pd
 
-from utils import wavesurfer_play
 
 sys.path.append("mtg-jamendo-dataset/scripts/")
 import commons
@@ -16,6 +15,92 @@ import commons
 
 def generate_uuid():
     st.session_state.user_uuid = str(uuid.uuid4())
+
+
+def wavesurfer_play(
+    tid: str,
+    gain: float = 1.0,
+) -> None:
+    """Play a track and print tags from its tid."""
+
+    jamendo_url = f"https://mp3d.jamendo.com/?trackid={tid}&format=mp32#t=0,120"
+
+    st.write("---")
+    st.write(f"**Track {tid}**")
+
+    html_code = f"""
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <script src="https://unpkg.com/wavesurfer.js"></script>
+    <style>
+        #waveform {{
+            width: 100%;
+            height: 128px;
+            margin: 0 auto;
+        }}
+        body {{
+            text-align: center;
+        }}
+    </style>
+</head>
+<body>
+    <div id="waveform"></div>
+
+    <script type="text/javascript">
+        document.addEventListener("DOMContentLoaded", function() {{
+            var wavesurfer = WaveSurfer.create({{
+                container: '#waveform',
+                waveColor: 'violet',
+                progressColor: 'purple',
+                height: 128,
+                barWidth: 2
+            }});
+
+            // Load audio from a URL and autoplay
+            wavesurfer.load('{jamendo_url}');
+            wavesurfer.setVolume({gain});
+            wavesurfer.on('ready', function() {{
+                wavesurfer.play();
+            }});
+
+            // Add play puse button
+            var playButton = document.createElement('button');
+
+            wavesurfer.on('play', function() {{
+                playButton.innerHTML = 'Pause';
+                playButton.style.position = 'center';
+                playButton.style.fontSize = '20px';
+                playButton.style.width = '100px';
+                playButton.style.backgroundColor = 'white';
+                playButton.style.border = '0px solid white';
+
+                playButton.onclick = function() {{
+                    wavesurfer.pause();
+                    playButton.innerHTML = 'Play';
+                    playButton.onclick = function() {{
+                        wavesurfer.play();
+                        playButton.innerHTML = 'Pause';
+                    }};
+                }};
+                document.getElementById('waveform').appendChild(playButton);
+            }});
+
+            // Add event listeners for mouse events to redirect focus
+            wavesurfer.on('interaction', function () {{
+                document.activeElement.blur();  // Remove focus from the current element
+                var focusElement = window.parent.document.querySelector('.main');
+                focusElement.focus();
+            }});
+        }});
+    </script>
+</body>
+</html>
+"""
+
+    # Embed the HTML code in the Streamlit app
+    st.components.v1.html(html_code, height=170)
 
 
 @st.cache_data
@@ -251,7 +336,7 @@ else:
     print(f"Track {tid} loudness: {loudness_db:.2f} dB, trim: {trim:.2f} dB")
     print(f"Playing track {tid} with gain {gain:.2f}")
 
-    wavesurfer_play(tid, tracks, autoplay=True, gain=gain)
+    wavesurfer_play(str(tid), gain=gain)
 
     n_rows = 2
     n_cols = len(choices) // n_rows
